django-shop/djangoshop/requirements.txt
Django==5.1.4
requests==2.32.3
pillow==11.1.0
django-storages==1.14.4
django-environ==0.11.2
django-celery-beat==2.7.0
boto3==1.35.95
celery==5.4.0
pytz==2024.2
redis==5.2.1
hiredis==3.1.0
psycopg2-binary==2.9.9